+++ travelBoard/src/main/java/edu/kh/travel/main/model/service/MainService.java
@@ -1,8 +1,18 @@
 package edu.kh.travel.main.model.service;
 
+import java.util.Map;
+
 import org.springframework.stereotype.Service;
+
+import edu.kh.travel.member.model.dto.Member;
 
 
 public interface MainService {
+
+	Member selectTelBirth(Map<String, Object> map);
+
+
+
+
 
 }
