+++ travelBoard/src/main/java/edu/kh/travel/member/model/mapper/MemberMapper.java
@@ -42,8 +42,16 @@ public interface MemberMapper {
 	public int checkNickname(String memberNickname);
 
 
-
-
 	public Member selectTelBirth(Map<String, Object> map);
+
+
+	
+
+
+
+
+
+
+
 
 }
