+++ travelBoard/src/main/java/edu/kh/travel/member/model/service/MemberService.java
@@ -38,8 +38,6 @@ public interface MemberService {
 
 
 
-	Member selectTelBirth(Map<String, Object> map);
-
