+++ travelBoard/src/main/java/edu/kh/travel/member/model/service/MemberServiceImpl.java
@@ -99,14 +99,9 @@ public class MemberServiceImpl implements MemberService{
 	
 	
 	
-	
 
 	
-	@Override
-	public Member selectTelBirth(Map<String, Object> map) {
-		
-		return mapper.selectTelBirth(map);
-	}
+
