+++ travelBoard/src/main/resources/mappers/main-mapper.xml
@@ -2,6 +2,16 @@
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd" >
 <mapper namespace="edu.kh.travel.main.model.mapper.MainMapper">
 
+<!-- 전화번호로 아이디 찾기  -->
+<select id="selectTelBirth" resultType ="Member">
+
+SELECT MEMBER_EMAIL
+FROM "MEMBER"
+WHERE MEMBER_DEL_FL = 'N'
+AND MEMBER_TEL = #{memberTel}
+AND MEMBER_BIRTH = #{memberBirth}
+
+</select>
