+++ travelBoard/src/main/resources/templates/main.html
@@ -24,7 +24,7 @@
                 <input type="checkbox" name="saveId">아이디 저장
             </label>
             <div id="loginText">
-                <a href="/member/signup">회원가입</a> | <a href="/member/searchId">ID 찾기</a> | <a href="/member/searchPw">PW 찾기</a>
+                <a href="/member/signup">회원가입</a> | <a href="/searchId">ID 찾기</a> | <a href="/searchPw">PW 찾기</a>
             </div>
             
             <button id="connectBoard">게시판 접속</button>
