+++ travelBoard/src/main/resources/templates/member/idResult.html
@@ -7,7 +7,9 @@
 </head>
 <body>
     <h1>아이디 찾기</h1>
-    <h1 th:text="${member.memberEmail}"></h1>
+
+    <h1><h1 th:text="${member.memberEmail}"></h1>
+   
     
     
 </body>
