+++ travelBoard/src/main/resources/templates/searchId.html
@@ -8,24 +8,25 @@
 </head>
 <body>
     <th:block th:replace="~{common/common}"></th:block>
-    <div>
+    <div class="idSearchMessage">
     <h1>아이디 찾기</h1>
     </div>
     <!-- 아이디 찾기하는데 필요한 요소 : 전화번호 입력  -->
     
-    <form action="/member/searchId" method="GET" id="searchIdForm" name="selectTelBirth">
+    <form action="/selectTelBirth" name="selectTelBirth">
     <div>
         <div id="telnumber">전화번호</div>  
-        <input type="text" id="memberTel" name = "memberTel" placeholder = "전화번호">
+        <input type="text" name = "inputTel" id="inputTel" placeholder = "전화번호">
     </div>
     <div>
         <div id="birthday">생년월일</div> 
-        <input type="text" id="memberBirth" name = "memberBirth" placeholder = "생년월일 (예 : 19901231)">
+        <input type="text" name = "inputBirth" id="inputBirth" placeholder = "생년월일 (예 : 19901231)">
     </div>
     <div>
         
-        <button>제출하기</button>
+        <button id="submit">제출하기</button>
     </div>
     </form>
 </body>
+<!-- "/member/searchId" -->
 </html>
